Implementacion CardiovascularMetrics

+++ src/main/java/healthcalc/CardiovascularMetrics.java
@@ -2,6 +2,6 @@ package healthcalc;
 
 public interface CardiovascularMetrics {
     
-    public double getIdealWeight(int height, char gender) throws Exception;
+    public float getIdealWeight(int height, char gender) throws Exception;
     
 }
